+++ tech1-framework-b2b-postgres-server/src/main/java/io/tech1/framework/b2b/postgres/server/tasks/ResetServerTask.java
@@ -1,0 +1,32 @@
+package io.tech1.framework.b2b.postgres.server.tasks;
+
+import io.tech1.framework.b2b.base.security.jwt.domain.jwt.JwtUser;
+import io.tech1.framework.b2b.base.security.jwt.tasks.AbstractSuperAdminResetServerTask;
+import io.tech1.framework.domain.system.reset_server.ResetServerStatus;
+import io.tech1.framework.incidents.events.publishers.IncidentPublisher;
+import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Getter
+@Component
+public class ResetServerTask extends AbstractSuperAdminResetServerTask {
+
+    private final ResetServerStatus status = new ResetServerStatus(1);
+
+    @Autowired
+    public ResetServerTask(
+            IncidentPublisher incidentPublisher
+    ) {
+        super(
+                incidentPublisher
+        );
+    }
+
+    @Override
+    public void resetOnServer(JwtUser user) {
+        this.status.complete();
+    }
+}