+++ RELEASE_NOTES.md
@@ -1,7 +1,8 @@
 # Release Notes
 
 ### v1.14
-- Completed abstract accessToken/refreshToken database security flow 
+- Add abstract reset server flow
+- Add abstract accessToken/refreshToken database security flow 
 - TimestampUtility: isBetweenInclusive
 - MongoDb: truncate/delete "tech1_user_sessions" (MongoDbUserSession: requestMetadata → metadata)
 - Users API: /user → /users + Postman